fix: example code compilation errors
- hello_qram.cpp: use std::vector<size_t> for memory
- grover_search.cpp: use state.basis_states for operators that expect std::vector<System>&
- Add _USE_MATH_DEFINES for Windows M_PI support

+++ examples/hello_qram.cpp
@@ -46,7 +46,7 @@ int main() {
     // Address 1 -> Value 3 (binary: 011)
     // Address 2 -> Value 7 (binary: 111)
     // Address 3 -> Value 1 (binary: 001)
-    std::vector<int> memory = {5, 3, 7, 1};
+    std::vector<size_t> memory = {5, 3, 7, 1};
     qram.set_memory(memory);
     
     std::cout << "Memory initialized:" << std::endl;
